server/requirements.txt
asgiref==3.3.1
Django==3.1.3
django-cors-headers==3.5.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
PyJWT==1.7.1
python-decouple==3.3
pytz==2020.4
sqlparse==0.4.1
